+++ optimus/renderer/_internal.py
@@ -132,9 +132,12 @@ from optimus.renderer.doc_event_renderer import (
 # focused future PR with proper coupling-graph design.
 from optimus.renderer.finding_enrichment import (
 	_GROUPING_SEVERITY_RANK,
+	_attach_drilldown_chains,
+	_find_node_in_tree,
 	_group_findings_by_root_cause,
 	_normalize_callsite,
 	_root_cause_key,
+	_walk_drilldown_chain,
 )
 
 # v0.12.12+: Line-Level Drilldown panel + per-function tables moved
@@ -1736,171 +1739,9 @@ _SQL_REDFLAG_FINDING_TYPES = frozenset({
 })
 
 
-def _find_node_in_tree(tree: dict, basename: str, function: str) -> dict | None:
-	"""Depth-first walk a pyinstrument call tree looking for a node that matches
-	``(basename(filename), function)``. Returns the first hit, or ``None``.
-
-	The basename match (rather than full path) survives bench-relative vs
-	absolute path differences between where the analyzer ran and where the
-	render is happening - same trick ``_build_line_drilldown_callsite_index``
-	uses.
-	"""
-	if not isinstance(tree, dict):
-		return None
-	want = (basename or "").strip()
-	want_fn = (function or "").strip()
-	if not want_fn:
-		return None
-	stack = [tree]
-	while stack:
-		node = stack.pop()
-		if not isinstance(node, dict):
-			continue
-		node_file = (node.get("filename") or "").rsplit("/", 1)[-1]
-		node_fn = (node.get("function") or "")
-		if node_fn == want_fn and (not want or node_file == want):
-			return node
-		children = node.get("children") or []
-		# DFS in-order via reverse-append so leftmost children pop first.
-		stack.extend(reversed(children))
-	return None
-
-
-def _walk_drilldown_chain(
-	tree: dict,
-	callsite: dict,
-	tracked_apps: tuple[str, ...] = (),
-	max_depth: int = 4,
-	signal_floor_pct: float = 10.0,
-) -> list[dict]:
-	"""Build a *Drill-down* chain below the finding's origin frame.
-
-	Given a per-action pyinstrument tree (dict form from
-	``analyzers/call_tree._walk_pyi_frame``) and a finding's callsite
-	(``{"filename": ..., "function": ...}``), locate the origin node then walk
-	hottest-child links downward until one of:
-
-	- the next child's filename is in framework code, OR
-	- depth reaches ``max_depth``, OR
-	- no children remain, OR
-	- the next child's ``cumulative_ms`` is below ``signal_floor_pct`` % of the
-	  origin's ``cumulative_ms`` (drops noisy near-leaf frames).
-
-	Returns a list of ``{filename, lineno, function, cumulative_ms,
-	pct_of_origin}`` dicts — one per level *below* the origin. The origin
-	itself is omitted (already rendered in the smoking-gun block).
-
-	Defensive: any malformed input → ``[]``.
-	"""
-	if not isinstance(tree, dict) or not isinstance(callsite, dict):
-		return []
-	filename = callsite.get("filename") or ""
-	function = callsite.get("function") or ""
-	if not function:
-		return []
-	from optimus.analyzers.base import is_framework_callsite
-
-	# If the finding's own callsite is already in framework code, the chain
-	# below would be even further from user-actionable code — skip.
-	if is_framework_callsite(filename, tracked_apps=tracked_apps or None):
-		return []
-
-	origin = _find_node_in_tree(tree, filename.rsplit("/", 1)[-1], function)
-	if origin is None:
-		return []
-
-	origin_ms = float(origin.get("cumulative_ms") or 0)
-	if origin_ms <= 0:
-		return []
-
-	floor_ms = origin_ms * (signal_floor_pct / 100.0)
-	chain: list[dict] = []
-	node = origin
-	for _ in range(max(0, max_depth)):
-		children = node.get("children") or []
-		if not children:
-			break
-		# Pick the hottest child by cumulative_ms, skipping synthetic
-		# "[other: N frames]" collapse nodes (v0.7.x: not shown in chains —
-		# you can't drill into a collapsed bucket).
-		hottest = max(
-			(c for c in children
-			 if isinstance(c, dict) and not _ct_is_other_frame(c.get("function"))),
-			key=lambda c: float(c.get("cumulative_ms") or 0),
-			default=None,
-		)
-		if hottest is None:
-			break
-		child_ms = float(hottest.get("cumulative_ms") or 0)
-		if child_ms < floor_ms:
-			break
-		child_file = hottest.get("filename") or ""
-		if is_framework_callsite(child_file, tracked_apps=tracked_apps or None):
-			break
-		chain.append({
-			"filename": child_file,
-			"lineno": int(hottest.get("lineno") or 0),
-			"function": hottest.get("function") or "",
-			"cumulative_ms": child_ms,
-			"pct_of_origin": int(round(child_ms / origin_ms * 100)) if origin_ms else 0,
-		})
-		node = hottest
-
-	return chain
-
-
-def _attach_drilldown_chains(findings, actions, tracked_apps: tuple[str, ...] = ()) -> None:
-	"""Walk each finding's representative call tree and attach a
-	``drilldown_chain`` to its ``technical_detail`` dict. Mutates findings in
-	place — same pattern as ``_attach_representative_callsites``.
-
-	Tree JSON parses are cached per ``action_idx`` so a session with several
-	findings on the same slow action only deserialises the tree once.
-	"""
-	if not findings or not actions:
-		return
-
-	# Index actions by their original ``idx`` so action_ref lookups survive the
-	# min_action_duration_ms filter (which preserves idx but reshapes the list).
-	actions_by_idx: dict[int, dict] = {}
-	for a in actions:
-		try:
-			actions_by_idx[int(a.get("idx"))] = a
-		except (TypeError, ValueError):
-			continue
-
-	tree_cache: dict[int, dict] = {}
-	for finding in findings:
-		detail = finding.get("technical_detail") or {}
-		callsite = detail.get("callsite") or {}
-		if not callsite.get("function"):
-			continue
-		ref = finding.get("action_ref")
-		if ref in (None, ""):
-			continue
-		try:
-			idx = int(ref)
-		except (TypeError, ValueError):
-			continue
-		action = actions_by_idx.get(idx)
-		if not action:
-			continue
-		if idx not in tree_cache:
-			try:
-				tree_cache[idx] = json.loads(action.get("call_tree_json") or "{}")
-			except (TypeError, ValueError):
-				tree_cache[idx] = {}
-		tree = tree_cache.get(idx) or {}
-		if not tree:
-			continue
-		chain = _walk_drilldown_chain(tree, callsite, tracked_apps=tracked_apps)
-		# v0.7.x: always attach the chain — even empty. The template
-		# distinguishes "key absent" (no callsite/action/tree, never
-		# attempted) from "empty list" (attempted but no eligible
-		# user-code descendants) to decide whether to render a "no
-		# deeper user-code frame" placeholder in place of the chain.
-		detail["drilldown_chain"] = chain
-		finding["technical_detail"] = detail
+# v0.12.19+: _find_node_in_tree / _walk_drilldown_chain /
+# _attach_drilldown_chains moved to optimus/renderer/finding_enrichment.py.
+# Re-imported at the top of this module so call sites resolve unchanged.
 
 
 def _attach_representative_callsites(findings, recordings, *, file_cache: dict | None = None) -> None:
